% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfgCost.R
\name{mfgCostUI}
\alias{mfgCostUI}
\title{产品分组UI}
\usage{
mfgCostUI(
  tabTitle = "产品生产成本",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = mfgCostFunc_left,
  func_right = mfgCostFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品分组UI
}
\examples{
mfgCostUI()
}
