% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mngrCost.R
\name{mngrCostUI}
\alias{mngrCostUI}
\title{产品分组UI}
\usage{
mngrCostUI(
  tabTitle = "产品管理成本",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = mngrCostFunc_left,
  func_right = mngrCostFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品分组UI
}
\examples{
mngrCostUI()
}
